Consolidate y-axis grid+label code into a function
We were doing essentially the same operations in 3 different places.
This makes it difficult to change how the y-axis grid lines and labels
are created, which also means it's error prone and likely to introduce
bugs.

Create one function that captures an expression to use for the ylim
argument, and uses left and right booleans to determine whether to add
left and/or right y-axis labels.

Use substitute() to create a language object containing the expression
we want, but with the ylim expression that was supplied as an argument.
Since substitute() returns a language object, we need to coerce it to
an expression before returning.